refactor: organize image, icons
