+++ @types/index.d.ts
@@ -14,3 +14,9 @@ interface IParentComponentProps {
 }
 
 declare module 'shortid';
+
+declare module '*.svg' {
+  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
+  const src: string;
+  export default src;
+}
