+++ src/assets/icons/dosopt_symbol.png
