+++ src/assets/mainLogo/base64_logo.ts
