+++ src/components/common/ScrollToTopButton/ScrollToTopButton.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { ReactComponent as UpArrow } from '@src/assets/icons/upArrow.svg';
+import { ReactComponent as UpArrow } from '@src/assets/icons/ic_up_arrow.svg';
 import { debounce } from '@src/utils/debounce';
 import cc from 'classcat';
 import styles from './scroll-to-top-button.module.scss';
