--- src/lib/types/image.d.ts
@@ -1,5 +1,0 @@
-declare module '*.svg' {
-  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
-  const src: string;
-  export default src;
-}
