+++ src/views/ProjectDetailPage/utils/getLinkNameAndSrcWithType.ts
@@ -1,9 +1,9 @@
 import appstore from '@src/assets/icons/appstore_icon.svg';
 import github from '@src/assets/icons/github_icon.svg';
 import googleplay from '@src/assets/icons/googleplay_icon.svg';
-import media from '@src/assets/icons/media_icon.svg';
-import website from '@src/assets/icons/website_icon.svg';
-import instagram from '@src/assets/logo/instagram.svg';
+import media from '@src/assets/icons/ic_media.svg';
+import website from '@src/assets/icons/ic_website.svg';
+import instagram from '@src/assets/icons/instagram.svg';
 import { LinkDetailType } from '../types';
 
 const LinkMap: Record<LinkDetailType, { name: string; src: string }> = {
