+++ src/views/ProjectPage/components/project/ProjectEnrollSection.tsx
@@ -1,4 +1,4 @@
-import { ReactComponent as RightArrow } from '@src/views/ProjectPage/assets/icon/arrow-right-24x24.svg';
+import { ReactComponent as RightArrow } from '@src/assets/icons/arrow_right_24x24.svg';
 import styles from './project-enroll.module.scss';
 
 export function ProjectEnrollSection() {
