+++ src/views/RulesPage/components/RulesList/RulesList.tsx
@@ -1,4 +1,4 @@
-import { RULES } from '@src/assets/sopt/rules';
+import { RULES } from '@src/lib/constants/rules';
 import CollapseLi from '../CollapseLi/CollapseLi';
 import * as S from './RulesList.style';
