+++ src/views/SopticlePage/components/Description/index.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { mainColor } from '@src/assets/sopt/mainColor';
+import { mainColor } from '@src/styles/colors';
 
 const Description = () => {
   return <Title>솝티클</Title>;
