+++ src/views/SponsorPage/components/CorporatePartner/index.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
+import { ReactComponent as ArrowLeft } from '@src/assets/icons/arrow_left_28x28.svg';
+import { ReactComponent as ArrowRight } from '@src/assets/icons/arrow_right_28x28.svg';
 import { useHorizontalScroll } from '@src/hooks/useHorizontalScroll';
-import { ReactComponent as ArrowLeft } from '@src/views/MainPage/assets/arrow-left-28x28.svg';
-import { ReactComponent as ArrowRight } from '@src/views/MainPage/assets/arrow-right-28x28.svg';
 import { corporatePartnerList } from './constants';
 import styles from './corporate-partner.module.scss';
